requirements.txt
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
arrow==1.2.3
asgiref==3.6.0
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1659291887007/work
Automat @ file:///home/conda/feedstock_root/build_artifacts/automat_1667331175863/work
bcrypt @ file:///D:/bld/bcrypt_1651441749287/work
binaryornot==0.4.4
brotlipy @ file:///D:/bld/brotlipy_1648854347500/work
certifi==2022.12.7
cffi @ file:///D:/bld/cffi_1656782940775/work
chardet==5.1.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
chess==1.9.3
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cookiecutter==2.1.1
cryptography @ file:///D:/bld/cryptography_1657174142266/work
cssselect @ file:///home/conda/feedstock_root/build_artifacts/cssselect_1666980406338/work
cssselect2==0.7.0
Django==4.1.4
django-appconf==1.0.5
django-browser-reload==1.6.0
django-compressor==4.1
django-compressor-toolkit==0.6.0
django-filer==2.2.3
django-js-asset==2.0.0
django-mptt==0.14.0
django-ninja==0.20.0
django-polymorphic==3.1.0
django-tailwind==3.4.0
easy-thumbnails==2.8.4
exceptiongroup==1.0.4
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1670289648023/work
h11==0.14.0
hyperlink @ file:///home/conda/feedstock_root/build_artifacts/hyperlink_1610092164190/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
incremental @ file:///home/conda/feedstock_root/build_artifacts/incremental_1665859450441/work
itemadapter @ file:///home/conda/feedstock_root/build_artifacts/itemadapter_1659518105741/work
itemloaders @ file:///home/conda/feedstock_root/build_artifacts/itemloaders_1661796597649/work
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath @ file:///home/conda/feedstock_root/build_artifacts/jmespath_1655568249366/work
lxml @ file:///C:/ci/lxml_1657527492694/work
MarkupSafe==2.1.1
numpy==1.24.0
outcome==1.2.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1670530880680/work
pandas==1.5.2
parsel @ file:///home/conda/feedstock_root/build_artifacts/parsel_1667392207259/work
Pillow==9.4.0
Protego @ file:///home/conda/feedstock_root/build_artifacts/protego_1647348122707/work
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==1.10.4
PyDispatcher==2.0.5
PyHamcrest @ file:///home/conda/feedstock_root/build_artifacts/pyhamcrest_1663745418692/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
PySocks @ file:///D:/bld/pysocks_1661604991356/work
pyspark==3.3.1
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7
pywin32==305.1
PyYAML==6.0
queuelib @ file:///home/conda/feedstock_root/build_artifacts/queuelib_1629994290383/work
rcssmin==1.1.0
reportlab==3.6.12
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
requests-file @ file:///home/conda/feedstock_root/build_artifacts/requests-file_1591227832914/work
rjsmin==1.2.0
Scrapy @ file:///C:/Windows/TEMP/abs_f50e21j997/croots/recipe/scrapy_1659598707153/work
selenium==4.7.2
service-identity==18.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
stockfish==3.28.0
svglib==1.4.1
text-unidecode==1.3
tinycss2==1.2.1
tldextract @ file:///home/conda/feedstock_root/build_artifacts/tldextract_1664943488272/work
trio==0.22.0
trio-websocket==0.9.2
Twisted @ file:///C:/Windows/Temp/abs_ccblv2rzfa/croots/recipe/twisted_1659592764512/work
twisted-iocpsupport @ file:///D:/bld/twisted-iocpsupport_1649418015679/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
tzdata==2022.7
Unidecode==1.1.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1669259737463/work
w3lib @ file:///home/conda/feedstock_root/build_artifacts/w3lib_1670599828182/work
webencodings==0.5.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wincertstore==0.2
wsproto==1.2.0
zope.interface @ file:///C:/ci_310/zope.interface_1642113633904/work
